+++ NEWS.md
@@ -2,6 +2,7 @@
 * plot_spec_sd default s.d. scaling factor is now 1.96 to aid 95% confidence
 interval visualisation.
 * Added subset option to mean_dyns function.
+* Added vline option to plot.
 
 # spant 2.15.0
 * An SNR estimate of zero in comb_coils is now a failure case.